requirements.txt
python-telegram-bot==12.4.2
Pillow==7.0.0
requests==2.23.0

